Rename solution package, modules and widgets
