+++ supervisely/__init__.py
@@ -272,7 +272,7 @@ from supervisely.task.progress import tqdm_sly
 import tqdm
 
 from supervisely import convert
-from supervisely import solutions
+from supervisely import solution
 
 _original_tqdm = tqdm.tqdm
