+++ supervisely/app/widgets/solution_card/style.css
