+++ supervisely/app/widgets/solution_graph/script.js
