+++ supervisely/app/widgets/solution_graph/style.css
