+++ supervisely/app/widgets/solution_graph/template.html
@@ -1,4 +1,4 @@
-<link rel="stylesheet" href="./sly/css/app/widgets/solutions_graph/style.css" />
+<link rel="stylesheet" href="./sly/css/app/widgets/solution_graph/style.css" />
 <div>
     <div :style="{height: data.{{{widget.widget_id}}}.height, width: data.{{{widget.widget_id}}}.width}" {% if
         widget._show_grid %} class="sly-flow-diagram dotted-grid" {% else %} class="sly-flow-diagram" {% endif %}>
