+++ supervisely/app/widgets/solution_project/style.css
