+++ supervisely/app/widgets/solution_project/template.html
@@ -1,4 +1,4 @@
-<link rel="stylesheet" href="./sly/css/app/widgets/solutions_project/style.css" />
+<link rel="stylesheet" href="./sly/css/app/widgets/solution_project/style.css" />
 
 <div class="solutions-node project-node" :style="{width: data.{{{widget.widget_id}}}.width}">
     <!-- badges -->
@@ -22,7 +22,7 @@
     <div v-if="data.{{{widget.widget_id}}}.disabled" class="wrapper disabled"></div>
 
     <!-- content -->
-    <div {% if widget._click_handled %} @click="post('/{{{widget.widget_id}}}/solutions_card_clicked_cb')" {% endif %}
+    <div {% if widget._click_handled %} @click="post('/{{{widget.widget_id}}}/solution_card_clicked_cb')" {% endif %}
         class="node-content-wrapper" :class="{
         'clickable': data.{{{widget.widget_id}}}.clickable,
         'node-loading': data.{{{widget.widget_id}}}.loading,
