+++ supervisely/solution/__init__.py
@@ -1,0 +1,5 @@
+from supervisely.solution.smart_sampling.smart_sampling import (
+    SamplingMode,
+    SamplingSettings,
+    SmartSampling,
+)
