+++ supervisely/solution/constants.py
