--- supervisely/solutions/__init__.py
@@ -1,5 +1,0 @@
-from supervisely.solutions.smart_sampling.smart_sampling import (
-    SamplingMode,
-    SamplingSettings,
-    SmartSampling,
-)
